# docs/archive/product-history/LobsterPark_Docs_V1_4/LobsterPark_OpenAPI_V1_4.yaml
openapi: 3.1.0
info:
  title: Lobster Park API
  version: 1.4.0
  description: >
    OpenAPI contract for Lobster Park. V1.4 adds Secrets CRUD endpoints,
    enhances alerts/notifications filters, fixes PatchInstanceRequest.name
    maxLength, and adds quotaJson to Tenant schema.
servers:
- url: https://lobster-park.internal
security:
- cookieAuth: []
paths:
...
  /api/v1/instances/{instanceId}:
    get:
      summary: Get instance detail
      operationId: getInstance
      parameters:
      - $ref: '#/components/parameters/InstanceId'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeInstance'
        '429':
          $ref: '#/components/responses/TooManyRequests'
    delete:
      summary: Delete instance
      operationId: deleteInstance
      parameters:
      - $ref: '#/components/parameters/InstanceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - confirmText
              properties:
                confirmText:
                  type: string
      responses:
        '202':
          description: accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeJobAccepted'
        '429':
          $ref: '#/components/responses/TooManyRequests'
    patch:
      summary: Update instance metadata
      operationId: patchInstance
      parameters:
      - $ref: '#/components/parameters/InstanceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchInstanceRequest'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeInstance'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
  /api/v1/instances/{instanceId}/nodes:
    get:
      summary: List instance nodes
      operationId: listInstanceNodes
      parameters:
      - $ref: '#/components/parameters/InstanceId'
      - $ref: '#/components/parameters/PageNo'
      - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopePagedNodes'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  # ── V1.4 NEW: Secrets CRUD ──────────────────────────────────────
  /api/v1/instances/{instanceId}/secrets:
    get:
      summary: List instance secrets (masked)
      operationId: listInstanceSecrets
      description: >
        Returns a paginated list of secrets for the given instance.
        Only masked previews are returned; cipher values are never exposed.
      parameters:
      - $ref: '#/components/parameters/InstanceId'
      - $ref: '#/components/parameters/PageNo'
      - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopePagedSecrets'
        '429':
          $ref: '#/components/responses/TooManyRequests'
    post:
      summary: Create instance secret
      operationId: createInstanceSecret
      description: >
        Encrypts the provided secret value with AES-256-GCM and stores it.
        The plain-text value is never persisted or returned.
      parameters:
      - $ref: '#/components/parameters/InstanceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSecretRequest'
      responses:
        '201':
          description: created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeSecret'
        '409':
          description: secret_key already exists for this instance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/instances/{instanceId}/secrets/{secretKey}:
    put:
      summary: Update instance secret
      operationId: updateInstanceSecret
      description: >
        Updates the secret value and/or expiration. If secretValue is provided,
        the cipher is re-encrypted and secret_version incremented.
      parameters:
      - $ref: '#/components/parameters/InstanceId'
      - $ref: '#/components/parameters/SecretKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSecretRequest'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeSecret'
        '404':
          description: secret not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
    delete:
      summary: Delete instance secret
      operationId: deleteInstanceSecret
      description: >
        Hard-deletes the secret. Fails if the secret is referenced by the
        current active config version.
      parameters:
      - $ref: '#/components/parameters/InstanceId'
      - $ref: '#/components/parameters/SecretKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteSecretRequest'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeBase'
        '409':
          description: secret is referenced by active config
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  # ── end Secrets CRUD ────────────────────────────────────────────
  /api/v1/nodes/pairing-requests:
    get:
      summary: List pairing requests
      operationId: listPairingRequests
      parameters:
      - $ref: '#/components/parameters/PageNo'
      - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopePagedPairingRequests'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/nodes/pairing-requests/{requestId}/approve:
    post:
      summary: Approve pairing request
      operationId: approvePairingRequest
      parameters:
      - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopePairingRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/nodes/pairing-requests/{requestId}/reject:
    post:
      summary: Reject pairing request
      operationId: rejectPairingRequest
      parameters:
      - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopePairingRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
  /api/v1/audits:
    get:
      summary: List audit logs
      operationId: listAudits
      parameters:
      - $ref: '#/components/parameters/PageNo'
      - $ref: '#/components/parameters/PageSize'
      - in: query
        name: tenantId
        schema:
          type: string
      - in: query
        name: instanceId
        schema:
          type: string
      - in: query
        name: actionType
        schema:
          type: string
      - in: query
        name: operatorId
        schema:
          type: string
      - in: query
        name: actionResult
        schema:
          type: string
          enum:
          - success
          - failed
      - in: query
        name: startTime
        schema:
          type: string
          format: date-time
      - in: query
        name: endTime
        schema:
          type: string
          format: date-time
      - in: query
        name: riskLevel
        schema:
          type: string
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopePagedAudits'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
  /api/v1/instances/{instanceId}/nodes/{nodeId}/detach:
    post:
      summary: Detach node from instance
      operationId: detachNode
      parameters:
      - $ref: '#/components/parameters/InstanceId'
      - $ref: '#/components/parameters/NodeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - confirmText
              properties:
                confirmText:
                  type: string
                reason:
                  type: string
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopePairingRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: lp_access
  parameters:
    PageNo:
      in: query
      name: pageNo
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      in: query
      name: pageSize
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    InstanceId:
      in: path
      name: instanceId
      required: true
      schema:
        type: string
...
    RequestId:
      in: path
      name: requestId
      required: true
      schema:
        type: string
...
    NodeId:
      in: path
      name: nodeId
      required: true
      schema:
        type: string
...
    SecretKey:
      in: path
      name: secretKey
      required: true
      schema:
        type: string
        pattern: '^[a-z0-9_]+$'
        maxLength: 64
      description: '[V1.4] Secret key identifier within an instance'
...
  schemas:
    EnvelopeBase:
      type: object
      properties:
        requestId:
          type: string
        code:
          type: integer
        message:
          type: string
    PagedMeta:
      type: object
      properties:
        pageNo:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
    InstanceLifecycleStatus:
      type: string
      enum:
      - draft
      - creating
      - create_failed
      - stopped
      - starting
      - running
      - unhealthy
      - updating
      - deleting
      - deleted
...
    PairingStatus:
      type: string
      enum:
      - pending
      - approved
      - rejected
      - expired
    NodeOnlineStatus:
      type: string
      enum:
      - online
      - offline
      - error
      - detached
...
    Instance:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        ownerUserId:
          type: string
        name:
          type: string
        specCode:
          type: string
          enum:
          - S
          - M
          - L
        runtimeVersion:
          type: string
        lifecycleStatus:
          $ref: '#/components/schemas/InstanceLifecycleStatus'
        healthStatus:
          type: string
        currentActiveVersionId:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        nodeCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PatchInstanceRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 64
          description: '[V1.4] maxLength aligned with CreateInstanceRequest (was 100)'
        description:
          type: string
          nullable: true
          maxLength: 500
    JobAccepted:
      type: object
      properties:
        jobId:
          type: string
        instanceId:
          type: string
...
    Node:
      type: object
      properties:
        id:
          type: string
        pairingStatus:
          $ref: '#/components/schemas/PairingStatus'
        onlineStatus:
          $ref: '#/components/schemas/NodeOnlineStatus'
        lastSeenAt:
          type: string
          format: date-time
          nullable: true
        tenantId:
          type: string
        boundInstanceId:
          type: string
          nullable: true
        metadataJson:
          type: object
          additionalProperties: true
        capabilitiesJson:
          type: array
          items:
            type: string
    PairingRequest:
      type: object
      properties:
        id:
          type: string
        instanceId:
          type: string
        pairingStatus:
          $ref: '#/components/schemas/PairingStatus'
        requestedAt:
          type: string
          format: date-time
        tenantId:
          type: string
        nodeFingerprint:
          type: string
        reviewedBy:
          type: string
          nullable: true
        reviewedAt:
          type: string
          format: date-time
          nullable: true
        reason:
          type: string
          nullable: true
...
    Audit:
      type: object
      properties:
        id:
          type: string
        actionType:
          type: string
        actionResult:
          type: string
        operatorUserId:
          type: string
        targetType:
          type: string
        targetId:
          type: string
        createdAt:
          type: string
          format: date-time
        tenantId:
          type: string
        summary:
          type: string
          nullable: true
        traceId:
          type: string
        riskLevel:
          type: string
        beforeJson:
          type: object
          additionalProperties: true
        afterJson:
          type: object
          additionalProperties: true
        metadataJson:
          type: object
          additionalProperties: true
    # ── V1.4 CHANGED: added quotaJson ──
...
    EnvelopeError:
      allOf:
      - $ref: '#/components/schemas/EnvelopeBase'
      - type: object
        properties:
          data:
            type: object
            additionalProperties: true
...
    SecretListItem:
      type: object
      description: '[V1.4] Masked secret item — cipher_value is never exposed'
      properties:
        id:
          type: string
          description: ULID with sec_ prefix
        secretKey:
          type: string
          description: Unique key within instance, e.g. "openai_api_key"
        maskedPreview:
          type: string
          description: Masked preview, e.g. "sk-****7890"
        secretVersion:
          type: integer
          description: Incremented on each value update
        expiresAt:
          type: string
          format: date-time
          nullable: true
          description: Optional expiration time for credential rotation tracking
        createdBy:
          type: string
        updatedBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateSecretRequest:
      type: object
      description: '[V1.4] Create a new secret for an instance'
      additionalProperties: false
      required:
      - secretKey
      - secretValue
      properties:
        secretKey:
          type: string
          pattern: '^[a-z0-9_]+$'
          maxLength: 64
          description: Unique identifier within the instance
        secretValue:
          type: string
          maxLength: 4096
          description: Plain-text secret value (encrypted at rest, never returned)
        expiresAt:
          type: string
          format: date-time
          nullable: true
          description: Optional expiration timestamp
    UpdateSecretRequest:
      type: object
      description: '[V1.4] Update secret value and/or expiration'
      additionalProperties: false
      minProperties: 1
      properties:
        secretValue:
          type: string
          maxLength: 4096
          description: New plain-text value (omit to keep current value)
        expiresAt:
          type: string
          format: date-time
          nullable: true
          description: New expiration (explicit null clears expiration)
    DeleteSecretRequest:
      type: object
      description: '[V1.4] Confirm secret deletion'
      additionalProperties: false
      required:
      - confirmText
      properties:
        confirmText:
          type: string
          enum:
          - DELETE
          description: Must be DELETE to confirm
    # ── end Secrets schemas ───────────────────────────────────────
    # ── Envelope wrappers ─────────────────────────────────────────
...
    EnvelopeInstance:
      allOf:
      - $ref: '#/components/schemas/EnvelopeBase'
      - type: object
        properties:
          data:
            $ref: '#/components/schemas/Instance'
...
    EnvelopeJobAccepted:
      allOf:
      - $ref: '#/components/schemas/EnvelopeBase'
      - type: object
        properties:
          data:
            $ref: '#/components/schemas/JobAccepted'
...
    EnvelopePagedNodes:
      allOf:
      - $ref: '#/components/schemas/EnvelopeBase'
      - type: object
        properties:
          data:
            allOf:
            - $ref: '#/components/schemas/PagedMeta'
            - type: object
              properties:
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/Node'
    EnvelopePagedPairingRequests:
      allOf:
      - $ref: '#/components/schemas/EnvelopeBase'
      - type: object
        properties:
          data:
            allOf:
            - $ref: '#/components/schemas/PagedMeta'
            - type: object
              properties:
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/PairingRequest'
    EnvelopePairingRequest:
      allOf:
      - $ref: '#/components/schemas/EnvelopeBase'
      - type: object
        properties:
          data:
            $ref: '#/components/schemas/PairingRequest'
...
    EnvelopePagedAudits:
      allOf:
      - $ref: '#/components/schemas/EnvelopeBase'
      - type: object
        properties:
          data:
            allOf:
            - $ref: '#/components/schemas/PagedMeta'
            - type: object
              properties:
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/Audit'
...
    EnvelopePagedSecrets:
      allOf:
      - $ref: '#/components/schemas/EnvelopeBase'
      - type: object
        properties:
          data:
            allOf:
            - $ref: '#/components/schemas/PagedMeta'
            - type: object
              properties:
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/SecretListItem'
    EnvelopeSecret:
      allOf:
      - $ref: '#/components/schemas/EnvelopeBase'
      - type: object
        properties:
          data:
            $ref: '#/components/schemas/SecretListItem'
  responses:
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/EnvelopeError'
...
  headers:
    RetryAfter:
      description: Seconds until the client may retry
      schema:
        type: integer